Move global_depolarizing_channel, amplitude_damping_channel, cosmic_ray_burst_profile out of mitigation.zne into dense_evolution.noise; add oscillating_p_eff, promoted from Discovery
