+++ dense_evolution/__init__.py
@@ -5,7 +5,8 @@ High-performance quantum statevector simulator optimized for NISQ circuits.
 from .backends.statevector import DenseSVSimulator
 from .circuits.parser import QASMParser, QASMCircuit
 from .circuits.compiler import QuantumTranspiler
-from .noise import NoiseModel, NoiseSpec
+from .noise import (NoiseModel, NoiseSpec, global_depolarizing_channel, amplitude_damping_channel,
+                     cosmic_ray_burst_profile, oscillating_p_eff)
 from .circuits.registry import QuantumHardwareRegistry
 from .circuits.gates import GATES, PARAMETRIC_GATES, GATE_IDS
 from .backends.chunk import Chunk
@@ -28,8 +29,7 @@ from .solvers.autodiff import circuit_to_energy_fn
 from .backends.mps import MPSSimulator
 from .mitigation.zne import (richardson_extrapolate, zero_noise_extrapolation, polynomial_extrapolate,
                           project_to_physical, uhlmann_fidelity, zne_density_matrix,
-                          jsd_predictive_zne_density_matrix, global_depolarizing_channel,
-                          amplitude_damping_channel, cosmic_ray_burst_profile,
+                          jsd_predictive_zne_density_matrix,
                           richardson_extrapolate_jit, zero_noise_extrapolation_jit,
                           polynomial_extrapolate_jit, uhlmann_fidelity_jit, zne_density_matrix_jit)
 from .circuits.diagram import plot_circuit
@@ -82,7 +82,7 @@ __all__ = [
     "richardson_extrapolate", "zero_noise_extrapolation", "polynomial_extrapolate",
     "project_to_physical", "uhlmann_fidelity", "zne_density_matrix",
     "jsd_predictive_zne_density_matrix", "global_depolarizing_channel", "amplitude_damping_channel",
-    "cosmic_ray_burst_profile",
+    "cosmic_ray_burst_profile", "oscillating_p_eff",
     "richardson_extrapolate_jit", "zero_noise_extrapolation_jit",
     "polynomial_extrapolate_jit", "uhlmann_fidelity_jit", "zne_density_matrix_jit",
     # Physics -- states, observables, entropy, fermions, QEC
